carp() when setting local patch fails
This code was silently broken for about 20 months because
patchlevel.h was detabified.

To avoid the same problem in the future: carp() when adding
the local patch to patchlevel.h fails we should keep an eye on the
logfiles.

+++ lib/Test/Smoke/Util.pm
@@ -680,8 +680,10 @@ sub set_local_patch {
     open PLIN,  "< $plh" or return 0;
     open PLOUT, "> $pln" or return 0;
     my $seen=0;
+    my $done=0;
     while ( <PLIN> ) {
         if ( /^(\s+),NULL/ and $seen ) {
+            $done++;
             while ( my $c = shift @descr ) {
                 print PLOUT qq{$1,"$c"\n};
            }
@@ -691,6 +693,12 @@ sub set_local_patch {
     }
     close PLIN;
     close PLOUT or return 0;
+
+    if ( not $done ) {
+        require Carp;
+        Carp::carp("Failed to update patchlevel.h. Content not as expected?");
+        return 0;
+    }
 
     -e $plb and 1 while unlink $plb;
     my $errno = "$!";
